Restructure project to Go standard layout with go install support
- Move main package to cmd/dnstap-filter/ for go install compatibility
- Split filters/ into internal/filter/ (node types + predicates) and
  internal/expression/ (expression parser)
- Rename filter files: ipfilter.go→ip.go, fqdnfilter.go→fqdn.go, etc.
- Update Makefile build target to ./cmd/dnstap-filter
- Fix .gitignore to use /dnstap-filter (root-only) to avoid ignoring cmd/
- Bump go version to 1.25.0 and update golang.org/x dependencies

Install: go install github.com/kimitoboku/dnstap-filter/cmd/dnstap-filter@latest
