+++ cmd/dnstap-filter/main_test.go
