+++ go.mod
@@ -1,12 +1,15 @@
 module github.com/kimitoboku/dnstap-filter
 
-go 1.14
+go 1.25.0
 
 require (
 	github.com/dnstap/golang-dnstap v0.4.0
 	github.com/miekg/dns v1.1.42
-	golang.org/x/crypto v0.0.0-20210506145944-38f3c27a63bf // indirect
-	golang.org/x/net v0.0.0-20210508051633-16afe75a6701 // indirect
-	golang.org/x/sys v0.0.0-20210507161434-a76c4d0a0096 // indirect
 	google.golang.org/protobuf v1.26.0
+)
+
+require (
+	github.com/farsightsec/golang-framestream v0.3.0 // indirect
+	golang.org/x/net v0.52.0 // indirect
+	golang.org/x/sys v0.42.0 // indirect
 )
